Update color_code.vbs
Small VBscript that color codes a label with predetermined color swatch folder stored in dropbox. Warning: Lots of ELSIF used.

Label was designed using NiceLabel, with locations G1 through G9 excluding G8. The same goes for H. H1 through H9 excluding H8.

Explanation:
If the label picking box letter is equals to "G" and the label picking box number is equals to a number 1 through 9 excluding 8 then the result is the color that you stored at that location in my case they colors are located in dropbox. Again, Same with H. I excluded 8 because we don't have that location available.